fix double shipping of orchestrator data in dca
