+++ releasenotes/notes/fix-cluster-agent-k8s-resource-dual-shipping-264514b8c4590153.yaml
@@ -1,0 +1,11 @@
+# Each section from every releasenote are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+fixes:
+  - |
+    Fix dual shipping for orchestrator resources in the cluster agent.
